server/requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
pymongo==4.10.1
python-dotenv==1.0.1
requests==2.32.3
httpx>=0.27.0
mcp
google-genai
supabase>=2.11.0
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.115.0